+++ Scripts/exec-mission-creator.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+#Author: Jesimar da Silva Arantes
+#Date: 28/11/2017
+#Last Update: 28/11/2017
+#Description: Script that executes the UAV-IFA software and manages the execution of the mission aiming at security aspects.
+#Descrição: Script que executa o software UAV-IFA e gerencia a execução da missão visando aspectos de segurança.
+
+cd ../UAV-Mission-Creator/
+java -jar dist/UAV-Mission-Creator.jar