+++ UAV-Generic/src/uav/generic/struct/constants/TypeAP.java
@@ -1,0 +1,12 @@
+package uav.generic.struct.constants;
+
+/**
+ * Classe que modela os tipos de pilotos automáticos possíveis de serem utilizados.
+ * @author Jesimar S. Arantes
+ */
+public class TypeAP {
+    
+    public static final String APM = "APM";
+    public static final String PIXHAWK = "PIXHAWK";
+    
+}