+++ UAV-Generic/src/uav/generic/struct/constants/TypeCC.java
@@ -1,0 +1,12 @@
+package uav.generic.struct.constants;
+
+/**
+ * Classe que modela os tipos de companion computer possíveis de serem utilizados.
+ * @author Jesimar S. Arantes
+ */
+public class TypeCC {
+    
+    public static final String EDISON = "EDISON";
+    public static final String RASPBERRY = "RASPBERRY";
+    
+}