+++ UAV-Generic/src/uav/generic/struct/constants/TypeOperationMode.java
@@ -7,8 +7,7 @@ package uav.generic.struct.constants;
 public class TypeOperationMode {
    
     public static final String SITL_LOCAL     = "SITL_LOCAL";
-    public static final String SITL_CC_EDISON = "SITL_CC_EDISON";
-    public static final String SITL_CC_RPI    = "SITL_CC_RPI";
+    public static final String SITL_CC        = "SITL_CC";
     public static final String REAL_FLIGHT    = "REAL_FLIGHT";
     
 }